requirements.txt
langgraph
langchain
langchain-groq
langchain-community
langchain-huggingface
python-dotenv
sentence-transformers
faiss-cpu
langchain-openai
langchain-embeddings